# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardAlertStatus = typeof DashboardAlertStatus[keyof typeof DashboardAlertStatus];


export const DashboardAlertStatus = {
  armed: 'armed',
  paused: 'paused',
  not_configured: 'not_configured',
} as const;

export interface Dashboard {
  walletBalance: number;
  dayPnl: number;
  dayPnlPercent: number;
  activeTrades: number;
  winRate: number;
  lastUpdated: string;
  alertStatus: DashboardAlertStatus;
}

export interface Quote {
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  currency: string;
  source: string;
  updatedAt: string;
}

export type ChannelStatus = typeof ChannelStatus[keyof typeof ChannelStatus];


export const ChannelStatus = {
  running: 'running',
  paused: 'paused',
  ready: 'ready',
} as const;

export interface Channel {
  id: string;
  name: string;
  description: string;
  icon: string;
  budget: number;
  active: boolean;
  status: ChannelStatus;
  accent: string;
}

export interface ChannelUpdate {
  /** @minimum 0 */
  budget?: number;
  active?: boolean;
}

export type TradeSide = typeof TradeSide[keyof typeof TradeSide];


export const TradeSide = {
  BUY: 'BUY',
  SELL: 'SELL',
} as const;

export type TradeStatus = typeof TradeStatus[keyof typeof TradeStatus];


export const TradeStatus = {
  watching: 'watching',
  protected: 'protected',
  executing: 'executing',
} as const;

export interface Trade {
  id: string;
  asset: string;
  channel: string;
  side: TradeSide;
  entryPrice: number;
  currentPrice: number;
  pnl: number;
  pnlPercent: number;
  status: TradeStatus;
  updatedAt: string;
}

export interface PerformancePoint {
  label: string;
  value: number;
  wins: number;
  trades: number;
}

export interface Performance {
  period: string;
  winRate: number;
  totalReturn: number;
  totalTrades: number;
  points: PerformancePoint[];
}

export interface WhatsappWebhookInput {
  Body?: string;
  From?: string;
  To?: string;
  message?: string;
}

export type GetPerformanceParams = {
period?: GetPerformancePeriod;
};

export type GetPerformancePeriod = typeof GetPerformancePeriod[keyof typeof GetPerformancePeriod];


export const GetPerformancePeriod = {
  week: 'week',
  fifteen_days: 'fifteen_days',
  month: 'month',
  year: 'year',
} as const;