# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelStatus } from './channelStatus';

export interface Channel {
  id: string;
  name: string;
  description: string;
  icon: string;
  budget: number;
  active: boolean;
  status: ChannelStatus;
  accent: string;
}