# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardAlertStatus } from './dashboardAlertStatus';

export interface Dashboard {
  walletBalance: number;
  dayPnl: number;
  dayPnlPercent: number;
  activeTrades: number;
  winRate: number;
  lastUpdated: Date;
  alertStatus: DashboardAlertStatus;
}