# lib/api-zod/src/generated/types/performance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PerformancePoint } from './performancePoint';

export interface Performance {
  period: string;
  winRate: number;
  totalReturn: number;
  totalTrades: number;
  points: PerformancePoint[];
}