# lib/api-zod/src/generated/types/trade.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeSide } from './tradeSide';
import type { TradeStatus } from './tradeStatus';

export interface Trade {
  id: string;
  asset: string;
  channel: string;
  side: TradeSide;
  entryPrice: number;
  currentPrice: number;
  pnl: number;
  pnlPercent: number;
  status: TradeStatus;
  updatedAt: Date;
}